# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: true
    type: object
info:
  contact: {}
  license: {}
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: 标签id
        in: query
        name: tag_id
        type: integer
      - description: access_token
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200, "data":{}, "msg":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 获取多个文章
      tags:
      - 文章管理
    post:
      parameters:
      - description: 标签id
        in: formData
        name: tag_id
        required: true
        type: integer
      - description: access_token
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200, "data":{}, "msg":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 新增文章
      tags:
      - 文章管理
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: 文章id
        in: path
        name: id
        required: true
        type: integer
      - description: access_token
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200, "data":{}, "msg":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 删除文章
      tags:
      - 文章管理
    get:
      parameters:
      - description: 文章id
        in: path
        name: id
        type: integer
      - description: access_token
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200, "data":{}, "msg":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 获取单个文章
      tags:
      - 文章管理
    put:
      parameters:
      - description: 文章id
        in: path
        name: id
        required: true
        type: integer
      - description: 标签id
        in: formData
        name: tag_id
        required: true
        type: integer
      - description: access_token
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200, "data":{}, "msg":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 修改文章
      tags:
      - 文章管理
  /api/v1/tags:
    get:
      parameters:
      - description: 标签名称
        in: query
        name: name
        type: string
      - description: 状态（0：禁用，1：正常）
        in: query
        name: state
        type: integer
      - description: access_token
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200, "data":{}, "msg":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 获取文章标签
      tags:
      - 标签管理
    post:
      parameters:
      - description: 标签名称
        in: formData
        name: name
        required: true
        type: string
      - description: access_token
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200, "data":{}, "msg":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 新增文章标签
      tags:
      - 标签管理
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 删除文章标签
      tags:
      - 标签管理
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: path
        name: token
        required: true
        type: string
      - description: 标签名称
        in: formData
        name: name
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 修改文章标签
      tags:
      - 标签管理
swagger: "2.0"